+++ packages/sdks/react-sdk/src/components/TenantProfile.tsx
@@ -22,7 +22,10 @@ const TenantProfileWC = lazy(async () => {
 });
 
 const TenantProfile = React.forwardRef<HTMLElement, TenantProfileProps>(
-  ({ logger, theme, debug, widgetId, styleId, tenant, multiSso, onReady }, ref) => {
+  (
+    { logger, theme, locale, debug, widgetId, styleId, tenant, onReady },
+    ref,
+  ) => {
     const [innerRef, setInnerRef] = useState(null);
 
     useImperativeHandle(ref, () => innerRef);
@@ -49,10 +52,10 @@ const TenantProfile = React.forwardRef<HTMLElement, TenantProfileProps>(
           baseCdnUrl={baseCdnUrl}
           styleId={styleId}
           tenant={tenant}
-          multiSso={multiSso}
           ref={setInnerRef}
           {...{
             'theme.attr': theme,
+            'locale.attr': locale,
             'debug.attr': debug,
             'styleId.attr': styleId,
             'tenant.attr': tenant,