+++ packages/widgets/tenant-profile-widget/src/lib/widget/api/apiPaths.ts
@@ -4,7 +4,6 @@ export const apiPaths = {
   },
   tenant: {
     details: '/v1/mgmt/tenant',
-    adminLinkSso: '/v1/mgmt/tenant/adminlinks/sso/authenticated',
-    ssoConfigurations: '/v1/mgmt/tenant/sso/configurations',
+    adminLinkSso: '/v1/mgmt/tenant/adminlinks/sso/authenticated/batch',
   },
 };