+++ packages/widgets/tenant-profile-widget/src/lib/widget/mixins/initMixin/initMixin.ts
@@ -1,8 +1,8 @@
 import { compose, createSingletonMixin } from '@descope/sdk-helpers';
 import { debuggerMixin, themeMixin } from '@descope/sdk-mixins';
 import { flowRedirectUrlMixin } from '../flowRedirectUrlMixin';
+import { initMultiSsoConfigurationsMixin } from './initComponentsMixins/initMultiSsoConfigurationsMixin';
 import { initTenantAdminLinkSSOMixin } from './initComponentsMixins/initTenantAdminLinkSSOMixin';
-import { initSsoConfigurationsListMixin } from './initComponentsMixins/initSsoConfigurationsListMixin';
 import { initTenantCustomAttributesMixin } from './initComponentsMixins/initTenantCustomAttributesMixin';
 import { initTenantEmailDomainsMixin } from './initComponentsMixins/initTenantEmailDomainsMixin';
 import { initTenantSSOExclusionsMixin } from './initComponentsMixins/initTenantSSOExclusionsMixin';
@@ -26,9 +26,9 @@ export const initMixin = createSingletonMixin(
       initTenantEnforceSSOMixin,
       initTenantCustomAttributesMixin,
       initTenantAdminLinkSSOMixin,
-      initSsoConfigurationsListMixin,
       initTenantPasswordPolicyUserAuthMethodMixin,
       initTenantSessionSettingsUserAuthMethodMixin,
+      initMultiSsoConfigurationsMixin,
     )(superclass) {
       async init() {
         await super.init?.();