+++ packages/widgets/tenant-profile-widget/src/lib/widget/mixins/stateManagementMixin.ts
@@ -5,14 +5,7 @@ import {
   initLifecycleMixin,
   loggerMixin,
 } from '@descope/sdk-mixins';
-import {
-  createSsoConfiguration,
-  deleteSsoConfiguration,
-  getMe,
-  getSsoConfigurations,
-  getTenant,
-  getTenantAdminLinkSSO,
-} from '../state/asyncActions';
+import { getMe, getTenant, getTenantAdminLinkSSO } from '../state/asyncActions';
 import { initialState } from '../state/initialState';
 import { apiMixin } from './apiMixin';
 
@@ -27,17 +20,11 @@ export const stateManagementMixin = createSingletonMixin(
           getMe.reducer(builder);
           getTenant.reducer(builder);
           getTenantAdminLinkSSO.reducer(builder);
-          getSsoConfigurations.reducer(builder);
-          createSsoConfiguration.reducer(builder);
-          deleteSsoConfiguration.reducer(builder);
         },
         asyncActions: {
           getMe: getMe.action,
           getTenant: getTenant.action,
           getTenantAdminLinkSSO: getTenantAdminLinkSSO.action,
-          getSsoConfigurations: getSsoConfigurations.action,
-          createSsoConfiguration: createSsoConfiguration.action,
-          deleteSsoConfiguration: deleteSsoConfiguration.action,
         },
       }),
       initLifecycleMixin,