+++ packages/widgets/tenant-profile-widget/src/lib/widget/state/asyncActions/index.ts
@@ -1,6 +1,3 @@
 export * from './getMe';
 export * from './getTenant';
 export * from './getTenantAdminLinkSSO';
-export * from './getSsoConfigurations';
-export * from './createSsoConfiguration';
-export * from './deleteSsoConfiguration';