+++ packages/sdks/web-component/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.67.1](https://github.com/descope/descope-js/compare/web-component-3.67.0...web-component-3.67.1) (2026-06-09)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.50.3`
 ## [3.67.0](https://github.com/descope/descope-js/compare/web-component-3.66.1...web-component-3.67.0) (2026-06-09)
 
 ### Dependency Updates