+++ packages/widgets/audit-management-widget/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.7.6](https://github.com/descope/descope-js/compare/audit-management-widget-0.7.5...audit-management-widget-0.7.6) (2026-06-09)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.50.3`
 ## [0.7.5](https://github.com/descope/descope-js/compare/audit-management-widget-0.7.4...audit-management-widget-0.7.5) (2026-06-09)
 
 ### Dependency Updates